# t3_log_backend/interfaces/openapi.running.downloaded.json
{"openapi":"3.1.0","info":{"title":"T3 Log Backend API","description":"API layer for T3 Log: mentorship workflows, tasks, and server-side notifications.\n\nEmail notifications are sent server-side via Resend. Frontend should never ship or access RESEND_API_KEY.","version":"0.1.0"},"paths":{"/":{"get":{"tags":["System"],"summary":"Health Check","description":"Health check endpoint.\n\nReturns:\n    JSON payload indicating service is alive.","operationId":"health_check","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/notifications/email/intern-submission":{"post":{"tags":["Notifications"],"summary":"Send email to mentor when intern submits a task update","description":"Looks up mentor and intern recipients via Supabase and sends a notification email via Resend. Requires RESEND_API_KEY, SUPABASE_URL, SUPABASE_ANON_KEY.","operationId":"send_intern_submission_email","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InternSubmissionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendEmailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notifications/email/mentor-update":{"post":{"tags":["Notifications"],"summary":"Send email to intern when mentor updates task status/remarks","description":"Looks up mentor and intern recipients via Supabase and sends a notification email via Resend. Requires RESEND_API_KEY, SUPABASE_URL, SUPABASE_ANON_KEY.","operationId":"send_mentor_update_email","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MentorUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendEmailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notifications/email/send":{"post":{"tags":["Email"],"summary":"Core server-side email send (Resend)","description":"Core server-side route to send a single email using Resend. Uses server-side RESEND_API_KEY only (frontend must never access it). Subject line should follow: 'T3 Log Update: [Action Name]'. Requires RESEND_API_KEY (or legacy RESEND_API-KEY) and RESEND_FROM_EMAIL.","operationId":"core_send_email","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoreSendEmailRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoreSendEmailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CoreSendEmailRequest":{"properties":{"to_email":{"type":"string","format":"email","title":"To Email","description":"Recipient email address."},"subject":{"type":"string","minLength":1,"title":"Subject","description":"Email subject."},"body":{"type":"string","minLength":1,"title":"Body","description":"Plain text / main body content."}},"type":"object","required":["to_email","subject","body"],"title":"CoreSendEmailRequest","description":"Request payload for sending an email via the server-side Resend integration."},"CoreSendEmailResponse":{"properties":{"resend_id":{"type":"string","title":"Resend Id","description":"Resend email id."},"to":{"items":{"type":"string","format":"email"},"type":"array","title":"To","description":"Recipient(s) the email was sent to."}},"type":"object","required":["resend_id","to"],"title":"CoreSendEmailResponse","description":"Response payload for the core email send endpoint."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InternSubmissionRequest":{"properties":{"intern_id":{"type":"string","title":"Intern Id","description":"Intern user id (Supabase auth uid or app user id)."},"mentor_id":{"type":"string","title":"Mentor Id","description":"Mentor user id (Supabase auth uid or app user id)."},"task_title":{"type":"string","title":"Task Title","description":"Task title."},"submission_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Submission Text","description":"Intern submission notes/body."}},"type":"object","required":["intern_id","mentor_id","task_title"],"title":"InternSubmissionRequest","description":"Request payload for notifying mentor of a new intern task submission."},"MentorUpdateRequest":{"properties":{"intern_id":{"type":"string","title":"Intern Id","description":"Intern user id (Supabase auth uid or app user id)."},"mentor_id":{"type":"string","title":"Mentor Id","description":"Mentor user id (Supabase auth uid or app user id)."},"task_title":{"type":"string","title":"Task Title","description":"Task title."},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status","description":"New status (e.g., Approved/Needs changes/Completed)."},"remarks":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remarks","description":"Mentor remarks/body."}},"type":"object","required":["intern_id","mentor_id","task_title"],"title":"MentorUpdateRequest","description":"Request payload for notifying intern of a mentor update."},"SendEmailResponse":{"properties":{"resend_id":{"type":"string","title":"Resend Id","description":"Resend email id."},"to":{"items":{"type":"string"},"type":"array","title":"To","description":"Resolved recipient emails."}},"type":"object","required":["resend_id","to"],"title":"SendEmailResponse","description":"Response indicating whether an email was successfully sent."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"System","description":"Health and system endpoints."},{"name":"Notifications","description":"Server-side notification endpoints (email, etc.)."},{"name":"Email","description":"Core server-side email send endpoints (Resend)."}]}